Maven assembly plugin use assembly:single goal
Replace the assembly:assembly goal with assembly:single goal in Maven
assembly plugin.
The assembly:assembly goal is deprecated. Maven 4 throws an error when
building ApiSummarizer.jar if the assembly:assembly goal is used.
Using assembly:single resolves this error on Maven 4 and also works on
Maven 3. We are using the maven package command so that tests are
executed.